pyproject.toml
[project]
name = "ocr-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marker-pdf>=1.10.2",
]
